[gdb] Move non-GDB specific functionality into its own class
